Connect Manage Judges Table to Backend
