+++ frontend/src/Admin/containers/JudgesTable.js
@@ -1,0 +1,11 @@
+import { graphql } from 'react-apollo'
+
+import JudgesQuery from '../queries/judges.graphql'
+import JudgesTable from '../components/JudgesTable'
+
+export default graphql(JudgesQuery, {
+  props: ({data: { users, loading }}) => ({
+    judges: users,
+    loading
+  })
+})(JudgesTable)
